Fontify Java constructor names and arglists
This fixes bug#63328.

* lisp/progmodes/cc-engine.el (c-forward-decl-or-cast-1): New
variable got-stmt-block.  After scanning an arglist, set
got-arglist to t.  When we have as yet no identifier,
got-arglist, and scan an open brace, test the "type" for being
the name of the enclosing class.
(c-directly-in-class-called-p): Test the two names for
equality, not merely one being the head of the other.